Added ExtensibleByAssociatedObject to allow any non-NSObject to use the associated methods easily.
